Style inline & XML Lint (#53)
Copy for email does not keep color.  We now do a copy of the color from the style sheet to the inline span element so that it remains.

Add option for xmldiff
--no-network
--no-entity

Missed getting the code to check XML lint correctly.  Now implemented w/ test cases.

* Update change logs and version

Update the change logs
Update the version numbers and requirements.txt
Fix test cases for xmldiff

* Fix type in change log

***** We still have a testing failure for pypy because of an error in lxml that now shows up.  Need to think about how to fix this at some point in the future.