Make word base diffs (#61)
* Make word diffs on a by word basis

Do word by word changing of differences rather than the old by character version

* Fix test cases

We are going to intentionally have 3 failures at this point.  The current code is wrong and the test cases are more correct however I want to get a drop out to Megan now

* Update version number

* Start the re-write to remove jquery

Update to use something different for doing the re-sizing of windows.  Also the goal is to remove all of the jquery code from the system.